pyproject.toml
[tool.poetry]
name = "pyTMBot"
version = "0.2.2-dev"
description = "Versatile Telegram bot designed for managing Docker containers, monitoring server status, and extending its functionality through a modular plugin system. The bot operates synchronously, simplifying deployment by eliminating the need for webhooks"
authors = ["Denis Rozhnovskiy <pytelemonbot@mail.ru>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/orenlab/pytmbot"

[tool.poetry.dependencies]
python = "^3.12"
emoji = "^2.9.0"
jinja2 = "^3.1.3"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
docker = "^7.0.0"
humanize = "^4.9.0"
psutil = "^6.0.0"
pyotp = "^2.9.0"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
pygal = "^3.0.5"
pytelegrambotapi = "^4.23.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
influxdb-client = "^1.46.0"
pypng = "^0.20220715.0"
qrcode = "7.4.2"
aiohttp = "==3.11.11"
pyoutlineapi = "^0.1.3"

[tool.poetry.plugins.outline]
pyoutlineapi = "^0.1.3"

[tool.poetry.dev-dependencies]
mypy = "^1.0.0"
packaging = "^24.1"
black = "^24.8.0"
isort = "^5.0.0"

[tool.poetry.extras]
pyoutlineapi = ["pyoutlineapi"]
click = ["click"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"